# public/swaggerdoc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen Service App Doc",
        "license": {
            "name": "TEDC"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "fe48bd38a69e79bda540d1b9227164b1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "User Name",
                                    "email": "user@gmail.com",
                                    "password": "123456",
                                    "password_confirmation": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/posts": {
            "get": {
                "tags": [
                    "public/posts"
                ],
                "summary": "Get Public posts",
                "operationId": "4ac4f9cd04624b3f1da30735c9c280c5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/public/posts/{id}": {
            "get": {
                "tags": [
                    "public/posts"
                ],
                "summary": "Get Public Posts by id",
                "operationId": "ae0c9e26b9886dbff22065924114bebb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}